adjust arrow angles by size
As suggested by tattsan, widen the arrow angle to help visibilty of
very non-slim arrows. The widening is done so that in that regime, the
tips of the arrow remain at the same distance from the line - this
requires resp. allows to shorten the arrow even more.

In the code, I left in many calculations with constants intentionally.
A good compiler will evaluate them as constants, and this should help
readability.

Suggested-By: SATO Tatsuya <tattsan@users.noreply.github.com>