WIP explain selection handles by floating tips
We use floating "tooltips" for CONTROL/SHIFT already (custom direction
cursor). Use the same principle for explaining the selection handles.

WIP because of readibility, and also because we might to do this for
selection deletion, as well - and maybe others?

Will be controversial anyway. WIP as a proof of concept, and because I
don't know how to transform icons into cursors.

OTOH, this does not localize well, just as CONTROL/SHIFT. Though we
could mark-up strings for l10n.