scale small arrow heads by line length (#4055)
* scale small arrow heads by line length

So far, the size of arrow heads is determined purely by the line width,
which is good for medium sized and long arrows. For very small arrows,
this leads to "interesting effects", especially for double headed
arrows.

To avoid these effects, scale the arrow head by the line length for
short arrows, i.e.: make the heads no larger than 0.4 * line length.
This suites double arrows as well, and besides, 0.4 is roughly
1 - (sqrt(5) - 1) / 2 which is the golden choice.

* scale arrow heads differently in more regimes

Depending on the ratio between linelength and thickness, different
choices of scalings show different artifacts. To avoid the worst ones,
choose more regimes for scaling (partly suggested by @tattsan).
For very "non-slim" arrows, we choose the cut-off depending on the
number of arrow heads so that the directionality remains visible when
shrinking down to a point. The resulting shape (arrowdist) is still a
linear function of (thickness, lineLength).

Suggested-By: SATO Tatsuya <tattsan@users.noreply.github.com>

* adjust arrow angles by size

As suggested by tattsan, widen the arrow angle to help visibilty of
very non-slim arrows. The widening is done so that in that regime, the
tips of the arrow remain at the same distance from the line - this
requires resp. allows to shorten the arrow even more.

In the code, I left in many calculations with constants intentionally.
A good compiler will evaluate them as constants, and this should help
readability.

Suggested-By: SATO Tatsuya <tattsan@users.noreply.github.com>