refactor storage of Sidebar-pages
Changes:
- std::list -> std::vector
- raw pointer -> std::unique_ptr

The advantage over the former approach of using a std::list is the
random access pattern. When using an std::vector, storing the active
page just via the index isn't bad as accessing the page is done in
constant time (random access container).

The downside of using a std::vector is that insertion might be slow (if
the vector needs to be copied), but as there are only a few pages and we
can preallocate space for the vector this is no big disadvantage.

Also we switched from using raw-pointers to smart-pointers
(std::unique_ptr) to benefit from the simplified destructor and use more
modern cpp.