ImageHandler: split insertImage up
createImage only creates the Image object from the provided file.
addImageToDocument inserts the image to the document
automaticScaling checks if the image at (x,y) with (width,height) fits onto the page and if not
scales it down with respect to the aspect ratio.

insertImage is still present (change is backwards-compatible) and just
redirects to the three new functions.