Make MacOS Finder open xopp-files
Finder uses Launch Services to identify Xournal++ as the
preferred app for xopp-files as declared in the Info.plist,
to launch it and send it an "Open Documents" event.
Gtk receives the event and uses g_application_open to
make the handler for the "open" signal react to it.
See gtkapplication-quartz.c.