Add explicit instantiations of the setSelectedPaperSize function template
Before, in some build configurations compilation failed because a
needed function instantiation for the setSelectedPaperSize function
template couldn't be found. This is because the setSelectedPaperSize is
in a cpp file and the compiler can't see the defintion in other
translation units. Because there are only two realistic values for the bool
changeComboBoxSelection template arg, I decided to create both options
explicitly. This seems easier than adding it to the header because there
you'd need to include new other headers to make it work and feasible
because only two values are possible.