Add page size selection to 'new page'-popover
The added menu elements in the popover allow for easily changing the page size of new pages (by selecting the page format that is wanted and clicking the 'new page' toolbar element) and existing pages (by using the buttons 'apply to current' and 'apply to all').

A couple variables are renamed to differentiate between setting the page type and the page size. Icons for the representing landscape and portrait orientation are added. When using the 'Configure Page Template'-dialogue and changing the default paper size in the dialogue, this change is reflected in the 'new page'-popover's paper format selection.