feat: Add recolor/inversion of drawing ares
This allows the user to not only have the UI in a dark mode (like it's
already possible) but also to get some sort of dark mode for the drawing
area.
We achieve this by inverting (+ some recoloring) the drawing area (and
optionally the previews shown in the sidebar as well).

Recoloring works in three steps:
1. Invert all colors
2. Scale the color spectrum to fit into the range of light and dark
   color (these are the two parameters of the recoloring)
3. Move the color spectrum to really be in that range.

This results in white areas being mapped to the *dark* color and black
areas being mapped to the *light* color. Everything in between is
interpolated to fit the color range.

All operations are applied at the RGB color channels individually (no
brightness/hue/saturartion calculation).

For the parametrization this adds four new settings (to the settings
dialog and the config file):
1. enable recoloring of the drawing area
2. enable recoloring of the previews in the sidebar
3. light color
4. dark color

Also if recoloring is enabled, we add an indicator on top of the color
buttons in the toolbar, of how the colors of the current palette will be
visible on the inverted drawing area.