QuickScreenshot: Do not block main thread while taking a screenshot
The way it was written, the UI froze every time a screenshot was taken using the plugin. This makes the plugin unusable in Xorg systems. Once the screenshot border is drawn to the screen Xorg sends an Expose event to the window, but it is not handled since the thread is too busy waiting for `maim`, so the screenshot is just a bunch of frozen borders.