Try to avoid shortening line strokes (#6622)
Lines drawn from point P to point Q may be recognized by the stroke
recognizer and drawn as a straight line. However, if there is a third
point R as a result of an accidental touch of the device (stylus against
the pad, or late release of the mouse's button) the stroke recognizer
may shorten the line in an unexpected way.