Recognize triangles
Copied the rectangle detection algorithm for triangles. The only
difference is the detection of orientation of the segments: I could not
make that work using an orientation test, and used a distance test.

When a triangle is detected, the strokes are simply straightened up and
no specific class of triangle is drawn. That is, the recognizer does not
discard, say, non-equilateral triangles, or triangles of other classes.