Intermediate Commit
I have no idea why it doesn't save the right pages, or why it saves other pages when I mark them, but it does save them. It also sometimes crashes when I try to undo changes, which is weird. It finds the right pages to write, but sometimes it doesn't.