Almost complete functionality
On a very complex document (many images, many strokes, etc.) of 140 pages, it went from 20 to 5-8 seconds; I think that's a win. It's still buggy because it doesn't take into account any redos the user might make, and it seems to pass the "pointer" of the changes to backup.xopp, which is in temp, and not to the ".xopp" document in target.