Fixed 2 bugs but introduced a Race Condition
Bug fix 1: When copying two pages, the changes from one were transferred to the other because I was copying the UID, which shouldn't be done.

Bug fix 2: The std::deque for the changes to be saved contained duplicate UIDs.

Race condition information: It appears that AutoSaveJob and h.saveTo in SaveJob.cpp are simultaneously accessing something they shouldn't be accessing. This needs to be fixed.