Updated reading the new document format
The changes only affect how pages are read. Specifically, it concerns
how layers are created and added to the page being parsed. Only one
`Layer` object is created per layer span; it is created at the first
page of every span. The pages that are not the first page of the span
will have a reference to the layer object created for the first page of
the span.

Consider a page 'p' being parsed. When the parser finds a of the page
field of the XML file finds a layer 'l' such that
- it spans over a single page, then 'l' is read from the file normally.
  A new `Layer` object is created.
- it spans over two or more pages:
  - If 'p' is the first page of the span, then the layer field of the
    XML file must contain the whole contents of 'l'. A new `Layer`
    object is created and added to the page.
  - If 'p' is not the first page, then the layer field is empty. Instead
    of creating a new `Layer` object we add to 'p' a reference to the
    `Layer` object that was created when parsing the page that contains
    the first reference to 'l'.